wip: introduce RecordDistribution records
This uses a similar trick as CommandapiRequestReader to use concrete
record value objects and fill them with the data stored in the record
value property. Otherwise, Java would not know which sub-class of
UnifiedRecordValue to use.

The rest of this code is inspired by IndexedRecord.setValue which helps
to understand how to write arbitrary bytes into a message pack object.