feat(stream-platform): check if key of record is generated on the current partition
Adds a method to the TypedRecord interface. This method will check if the record key was generated on the current partition. If this is the case it will return true, otherwise false.

A key could be generated on a different partition in the event of a command distribution. The key will be generated on partition X and distributed to partition Y.