feat(engine): acknowledge command in CommandDistributionBehavior
Adds a method to acknowledge commands in the CommandDistributionBehavior. This method take a distribution key. This key was generated on the distributing partition. This means we can decode the partitionId of the distributing partition from this key. We will send an ACKNOWLEDGE command to this partitionId to notify it that the the distribution has been completed on the current partition.