feat(engine): handle process Deployment CREATE on distributed partitions
Handles a Deployment CREATE command on partitions where the command has been distributed to. DMN is ignored for now. It will write a CREATE event for the processes. Once that's done it will send an acknowledgement back to the distributing partition. Finally it writes a Deployment CREATED event to finalize the deployment lifecycle for this partition.