fix(engine): return null when ValueType is NULL_VAL
The command value is not always set. This happens, for example, when we write the DISTRIBUTING events. In order to reduce the message size we leave out the commandValue. In these instances the ValueType will be NULL_VAL. In order to not run into exception we must add a check in the getCommandValue method to return null if the ValueType is NULL_VAL.

We also don't need to check if the ValueType is set. As the valueTypeProperty defines a default of NULL_VAL this property will always have a value.