feat(protocol): prevent limiting of CommandDistribution ACKNOWLEDGE commands
The white list we have here is used to allow some command to be processed, even though we have reached hit our limit.
Sometimes there is a process which is spamming commands. Performing a new deployment to fix the process could be a viable solution for the spam. In order to allow this to happen we have to make sure we don't reject the commands related to distribution.