feat(engine): allow appending specific versions of events
This adds an overloaded method to appendFollowUpEvents with a specific
record version. The event's metadata will contain the specified
recordVersion.

In a future commit, this will allow event appliers to become
recordVersion specific. That is necessary to allow different state
changes to result from applying a newer version of an event, while
making sure that each already written event continues to produce the
same state changes when it is replayed.