feat(engine): support versioned event appliers
Each event applier should produce the same state changes for each
appended event on replay, irrespective of the Zeebe version. If changes
need to be made to an event applier then this would lead to different
state changes on replay for events that have already been appended to the
log.

We can already append events for a specific record version. By using
that version to determine the event applier, we can support versioned
event appliers.

When we need to fix a bug, we can only fix it on events that haven't yet
been appended to the log. For events that are newly appended, we can
choose to append a new version of that event and create a new event
applier with the bug fix.

We can also use this to stay backward compatible and continue to write
previous versions of events in places where we need to support both
versions of an event simultaneously. For example, for the backward
compatibility of Deployment Distribution while we replace this with
Generalized Record Distribution.