feat(broker): add support for gossiping ClusterTopology
Initial ClusterTopology is gossiped by each member. Any update
received via gossip is also reflected in the local state. To make
progress faster, we have used json for serialization. This is
temporary until we decide the serializatiion format.

In this commit, we only added support for it. This is not yet
integrated with the broker.