wip: test: two event triggers where one is expected
This adds a test case where the ProcessProcessor's onActivate is tested
(although indirectly, because we try not to test implementation).

The idea is that a process instance started for a message start event
should peek at a unique event trigger (the one that triggered this
message event). However, that may not be the case.

This adds a test case where we show what happens when an additional
Event Trigger exists in the state.