test: use IdentityInterceptor in GatewayTest
The GatewayTest sets up its own Gateway to run the tests against. This
gateway did not yet have the IdentityInterceptor setup. So none of the
tests (except for the IdentityInterceptorTest which are unit tests)
actually passed through the IdentityInterceptor.

To be able to verify that the authorized tenants are set correctly on
the broker request, we'll need to be able to control the Identity sdk
used by the IdentityInterceptor as well.

This commit adds the IdentityInterceptor to the StubbedGateway, allows
configuring the GatewayCfg for the test, and control the mocked Identity
sdk used by the IdentityInterceptor.

It also makes sure a fake bearer token (Exactly: "Bearer token") is set
in the Authorization header of each client call. This is necessary
because otherwise the IdentityInterceptor rejects each incoming request
as 'missing Authorization header'.