test: ensure broker request has auth with tenants
This test ensures that the Authorization of a broker request contains
the authorized tenants provided by the IdentityInterceptor.

As the client requires a concrete request, I've selected the
DeployResourceRequest. Note that the code that is tested is request type
agnostic.

This test case not verify tenant access, but just that the information
is passed on to the broker. Verifying the tenant access (i.e. is a
client allowed to make a request for tenant-a when it is authorized for
tenant-b only?) must be done per request type, because the specific
tenant id is stored in each specific request. For example, the gateway
cannot check this for CompleteJob because the request will not refer to
a specific tenant id.