feat: reject request for unauthorized tenant
When a request is made for a specific tenant, but the client is not
authorized for that tenant, then the request must be rejected.

We check whether the authorized tenants contains the specified tenant id
from the request.

As the tenant id is mapped for each specific request type, tests should
be added for each request type as well.