fix: only subscribe to events if no other incident
This resolves a critical bug for call activities, where the activity is
subscribed to its events a second time when resolving any incident
created after subscribing to those events.

For example, a call activity with a message boundary event may subscribe
to that event and then still fail because the process could not be
found, or the process has not a none start event. On resolving the
incident, the activation of the call activity is continued which would
result in the subscribing the activity to the events again.

By first evaluating all expressions and checking all pre-conditions,
we can safely create the incident without any subscriptions. Note that
the call to subscribeToEvents itself first evaluates all event
expressions, before creating the subscriptions. So that part is safe as
well.

This is only a patch to cover up a problem. Eventually, we need a safer
way of creating incidents without any other effects.