fix(topology): apply operations consecutively on same member
Previously, apply pending operation was only triggered when receiving a new
topology via gossip. When two operations in sequence apply to the same member,
the second one was never triggered because the member doesn't receive gossip from
itself. As a result, topology changes are stuck forever.

To fix this, we now trigger applying changes immediately after the current changes
are applied. It is scheduled to the executor so as not to block the actor.