test: verify that message subscription is migrated
We added a test case where we assert the message subscription migrated events. We added assertions
for the changes in process definition fields that should be updated. Also, we add an assertion for
correlation key that we expect it should not change during the migration. The reason for that is we
do not support re-evaluating correlation key expression in this phase of Process Instance Migration.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>