feat: adjust target definition of proc-msg-sub
When migrating the process message subscription record, we need to
adjust some of its properties:
- bpmn process id
- element id (catch event element id)

Other properties (like message name and correlation key) cannot change
yet. We'll need to add those later.

One thing to note could be that the process message subscription record
is not stored safely if there are multiple, so we may need to copy them.
We'll add a test case for this later.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>